requirements.txt
arrow==1.3.0
asgiref==3.8.1
binaryornot==0.4.4
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
cookiecutter==2.6.0
coverage==7.8.1
diff-match-patch==20241021
Django==5.2.1
django-cleanup==9.0.0
django-environ==0.12.0
django-import-export==4.3.7
django-iso3166==10.1.1
django-pytest==0.2.0
django-tailwind==4.0.1
django-widget-tweaks==1.5.0
et_xmlfile==2.0.0
idna==3.10
iniconfig==2.1.0
iso3166==2.1.1
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
openpyxl==3.1.5
packaging==25.0
pluggy==1.6.0
Pygments==2.19.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
setuptools==80.8.0
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
text-unidecode==1.3
types-python-dateutil==2.9.0.20250516
tzdata==2025.2
urllib3==2.4.0
